1.쿠폰 관련 오류 수정 2.쿠폰 관리 페이지 수정 3.관리자 페이지, 메인페이지 스타일 수정

+++ src/main/java/com/soloProject/myTrip/dto/CouponSearchDto.java
@@ -9,6 +9,6 @@ import lombok.Setter;
 public class CouponSearchDto {
     private String searchDateType; // 검색 기간 타입
     private String couponType; // 쿠폰 타입 (AMOUNT/PERCENTAGE)
-    private Long minAmount; // 최소 구매 금액
+    private int minAmount; // 최소 구매 금액
     private String searchQuery; // 검색어 (쿠폰 설명)
 }
